# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/v1/plants": {
      "get": {
        "description": "Get all plants for the authenticated user",
        "summary": "Get all plants",
        "tags": [
          "Plant"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "name",
                "species",
                "location",
                "planted_date",
                "sort_order"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "species",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "plantedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "plantedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "operationId": "getPlants",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 100
                          },
                          "species": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 100
                          },
                          "location": {
                            "type": "string",
                            "maxLength": 255,
                            "nullable": true
                          },
                          "plantedDate": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "notes": {
                            "type": "string",
                            "maxLength": 1000,
                            "nullable": true
                          },
                          "metadata": {
                            "type": "object",
                            "properties": {
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "wateringFrequency": {
                                "type": "string"
                              },
                              "lastWateredAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "sunlightLevel": {
                                "type": "string"
                              },
                              "soilType": {
                                "type": "string"
                              },
                              "potSizeCm": {
                                "type": "number"
                              },
                              "fertilizerType": {
                                "type": "string"
                              },
                              "lastFertilizedAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "lastWeatherSnapshotId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "averageTempC": {
                                "type": "number"
                              },
                              "averageSunshineHrs": {
                                "type": "number"
                              },
                              "healthStatus": {
                                "type": "string"
                              },
                              "growthStage": {
                                "type": "string"
                              },
                              "heightCm": {
                                "type": "number"
                              },
                              "colorTag": {
                                "type": "string"
                              },
                              "imageUrl": {
                                "type": "string",
                                "format": "uri"
                              },
                              "emojiIcon": {
                                "type": "string"
                              },
                              "aiInsightSummary": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": true,
                            "nullable": true
                          },
                          "sortOrder": {
                            "type": "integer"
                          },
                          "observations": {
                            "type": "array",
                            "items": {
                              "nullable": true
                            }
                          }
                        },
                        "required": [
                          "id",
                          "createdAt",
                          "updatedAt",
                          "userId",
                          "name",
                          "species",
                          "location",
                          "plantedDate",
                          "notes",
                          "metadata",
                          "sortOrder",
                          "observations"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit",
                    "totalPages"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Create a new plant for the authenticated user",
        "summary": "Create a new plant",
        "tags": [
          "Plant"
        ],
        "parameters": [],
        "operationId": "createPlant",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "species": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "plantedDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "metadata": {
                    "type": "object",
                    "properties": {
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "wateringFrequency": {
                        "type": "string"
                      },
                      "lastWateredAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "sunlightLevel": {
                        "type": "string"
                      },
                      "soilType": {
                        "type": "string"
                      },
                      "potSizeCm": {
                        "type": "number"
                      },
                      "fertilizerType": {
                        "type": "string"
                      },
                      "lastFertilizedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "lastWeatherSnapshotId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "averageTempC": {
                        "type": "number"
                      },
                      "averageSunshineHrs": {
                        "type": "number"
                      },
                      "healthStatus": {
                        "type": "string"
                      },
                      "growthStage": {
                        "type": "string"
                      },
                      "heightCm": {
                        "type": "number"
                      },
                      "colorTag": {
                        "type": "string"
                      },
                      "imageUrl": {
                        "type": "string",
                        "format": "uri"
                      },
                      "emojiIcon": {
                        "type": "string"
                      },
                      "aiInsightSummary": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": true
                  }
                },
                "required": [
                  "name",
                  "species"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "userId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 100
                        },
                        "species": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 100
                        },
                        "location": {
                          "type": "string",
                          "maxLength": 255,
                          "nullable": true
                        },
                        "plantedDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "notes": {
                          "type": "string",
                          "maxLength": 1000,
                          "nullable": true
                        },
                        "metadata": {
                          "type": "object",
                          "properties": {
                            "tags": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "wateringFrequency": {
                              "type": "string"
                            },
                            "lastWateredAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "sunlightLevel": {
                              "type": "string"
                            },
                            "soilType": {
                              "type": "string"
                            },
                            "potSizeCm": {
                              "type": "number"
                            },
                            "fertilizerType": {
                              "type": "string"
                            },
                            "lastFertilizedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "lastWeatherSnapshotId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "averageTempC": {
                              "type": "number"
                            },
                            "averageSunshineHrs": {
                              "type": "number"
                            },
                            "healthStatus": {
                              "type": "string"
                            },
                            "growthStage": {
                              "type": "string"
                            },
                            "heightCm": {
                              "type": "number"
                            },
                            "colorTag": {
                              "type": "string"
                            },
                            "imageUrl": {
                              "type": "string",
                              "format": "uri"
                            },
                            "emojiIcon": {
                              "type": "string"
                            },
                            "aiInsightSummary": {
                              "type": "string"
                            }
                          },
                          "additionalProperties": true,
                          "nullable": true
                        },
                        "sortOrder": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "userId",
                        "name",
                        "species",
                        "location",
                        "plantedDate",
                        "notes",
                        "metadata",
                        "sortOrder"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "gardenjournal REST API - Documentation",
    "description": "gardenjournal REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}